Week 3. Collinear points. Refactoring.

+++ src/main/java/week_three/Point.java
@@ -4,13 +4,14 @@ import edu.princeton.cs.algs4.StdDraw;
 import java.util.Comparator;
 
 public class Point implements Comparable<Point> {
-  private final double x;
-  private final double y;
+  private final int x;
+  private final int y;
 
-  public Point(double x, double y) {
+  public Point(int x, int y) {
     this.x = x;
     this.y = y;
   }
+
 
   public void draw() {
     StdDraw.point(x, y);
